Impliment common HTTP calls via basic cli
The basic usage is ./call-buddy <call-type> <url> [content-type],
where the call-type is either POST, GET, HEAD, PUT or DELETE. If the
call-type is not a GET, HEAD or DELETE then stdin is used as the
request body and the content-type is text/plain, unless specified with
the optional content-type argument.

This is obviously a long way away from the UI and the full functionality
of call buddy, but the code showcases the basic code for making HTTP
calls.