Create launchpad wrapper program that execs call-buddy remotely via ssh
This commit introduces a new command called "launchpad", which wraps
around the call-buddy command and allows for the command to be launched
on any host via ssh. This is done via the -t flag, which specifies the
target. The wrapper also automatically detects the architecture of the
target host and uploads the right binary type. You can also override the
architecture via the -a flag.

There are some limitations with this. The launchpad program must be run
from the telephono/ directory after a 'make':

./cmd/launchpad/launchpad

Furthermore, the detection and looking for the binaries is not very
nice. It's hardcoded and it won't work with things like $PATH (e.g. if
the call-buddy binary is installed in /usr/bin and referenced by the
$PATH, then the program would assume that the build/ directory is in the
cwd).

Pretty Neat :^)